InputConverterAttribute to support interfaces
I would like the ability to convert a function input to an instance of an interface using `[InputConverter(...)]`. The only thing that seems to prevent this is the `AttributeUsage` declaration on the `InputConverterAttribute` class. I've added `AttributeTargets.Interface` to fix this.  I manually tested this with my Durable Functions bindings and it works great.