Ensure setup-e2e-tests.ps1 requires PowerShell 6+ (#1396)
Running the script as is in a PS5 console on Windows blindly downloads the linux build. Which is... unexpected ;-)

Indeed, $IsWindows and friends are only available in PS6+.

The proposed change requires PS6 as a minimum to be upfront with expectations rather than having the caller try and troubleshoot the problem by himself/herself.