Use name of function instead of class for input/output example (#3212)
The example function with input/output is used standalone on https://learn.microsoft.com/en-us/azure/azure-functions/functions-bindings-storage-queue-trigger . When copying this the code does not compile because it uses the class name, which is not included in the example (doc snippet), as function name.

Renaming the function to something more explicit, and using that as function name, should improve the example.